[tests] prepare for PHPUnit 12 stub upgrade

+++ app/bundles/LeadBundle/Tests/Entity/ImportTest.php
@@ -88,7 +88,7 @@ class ImportTest extends StandardImportTestHelper
         $import->increaseIgnoredCount();
         $import->increaseIgnoredCount();
 
-        $expectedCount = (int) (2 + $expectedCount);
+        $expectedCount = 2 + $expectedCount;
         $this->assertSame($expectedCount, $import->getProcessedRows()); // @phpstan-ignore argument.unresolvableType (I don't see anything wrong)
     }
